refactor: split OWASP/arch/compliance builders into dedicated modules (Option B)
Implements .omc/plans/dashboard-standards-split.md (Option B): three
dedicated modules + re-export shims kept in dashboard_html_sections.py
for backwards compatibility.

New modules:
- dashboard_html_owasp.py (84 lines)  - _build_owasp_html
- dashboard_html_arch.py (97 lines)   - _build_arch_html
- dashboard_html_compliance.py (77 lines) - _build_compliance_html

Changes:
- dashboard_html_sections.py 965 -> 769 lines (-196). Function bodies
  removed; re-imports added so existing import sites in dashboard-gen.py
  and scanner/tests/test_dashboard_builders_unit.py keep working.
- Dropped now-unused names from sections.py imports
  (OWASP_2025, OWASP_LLM_2025, OWASP_TO_ARCH, ARCH_DOMAINS,
  COMPLIANCE_FRAMEWORKS, comp_slug).
- test_dashboard_html_modules.py: smoke-imports the three new modules.

Verified:
- python3 -m py_compile on all four files - clean
- pytest scanner/tests/ -q -> 232 passed, 205 subtests passed
- Dashboard HTML byte-identical: 225080 == 225080 bytes
- diff: 8 lines (4 hunks - CSP nonce + timestamps only)
